+++ CHANGELOG.md
@@ -38,6 +38,7 @@ According to [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) , the `Unr
 
 -   [#193](https://github.com/ARGOeu/argo-accounting/pull/193) - ACC-274 Update Group ID Path Name to Resolve Conflict with User ID Path.
 -   [#194](https://github.com/ARGOeu/argo-accounting/pull/194) - ACC-275 Fix Search of Metrics by Start or End Date.
+-   [#195](https://github.com/ARGOeu/argo-accounting/pull/195) - ACC-286 Project entry "lost" when updated.
 
 ## 1.5.1 - 2024-07-24